pyproject.toml
[project]
name = "aurelia"
version = "1.0.0"
description = "AURELIA Health Coach - AI-powered health optimization with evidence-based recommendations"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agents>=1.4.0",
    "beautifulsoup4>=4.14.2",
    "ddgs>=9.9.1",
    "fastapi>=0.121.3",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "torch>=2.9.1",
    "transformers>=4.57.1",
    "uvicorn>=0.38.0",
]

[tool.setuptools]
packages = ["src.aurelia"]

[tool.setuptools.package-dir]
"" = "src"

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
